+++ packages/Icon/src/index.ts
@@ -2,5 +2,5 @@ import createIconComponent from './createIconComponent';
 import glyphs from './glyphs';
 
 export { glyphs, createIconComponent };
-export { Size, IconProps } from './createIconComponent';
+export { Size } from './createIconComponent';
 export default createIconComponent(glyphs);
